edited quotes route

+++ app/routes/quotes.py
@@ -62,7 +62,7 @@ def create_new_book_quote(book_id):
         if not content or content == '':
             return jsonify({'error': 'Missing required fields.'}), HTTP_400_BAD_REQUEST
         
-        # get the book ID
+        # check the book if it exists
         book = Book.query.filter_by(id=book_id).first()
         if not book:
             return jsonify({'error': 'Book not found.'}), HTTP_404_NOT_FOUND
@@ -100,7 +100,7 @@ def get_specific_quote(quote_id):
     ), HTTP_200_OK
 
 # delete a specific post
-@user_quotes.route('/delete/<int:quote_id>', methods=['DELETE', 'GET'])
+@user_quotes.route('/delete/<int:quote_id>', methods=['DELETE'])
 @jwt_required()
 def delete_quote(quote_id):
     userId = get_jwt_identity()
@@ -112,8 +112,10 @@ def delete_quote(quote_id):
     
     if not quote:
         return jsonify({'error': 'Quote not available.'}), HTTP_404_NOT_FOUND
-    db.session.delete(quote)
-    db.session.commit()
 
-    return jsonify({'message': 'Quote deleted!'}), HTTP_200_OK
+    if request.method == "DELETE":
+        db.session.delete(quote)
+        db.session.commit()
+
+        return jsonify({'message': 'Quote deleted successfully!'}), HTTP_200_OK
